add _failing_ tests which indicate the current state of development
The index is indexing all revisions, so we need to make sure when
we return results we only return the HEAD. There's a variety of ways
to do this, and none of the clever ways are working for me right now
so I'm retiring for the evening to revisit tomorrow. The failing tests
help me to park on a slope.