Further work to get search and index_query working.
The major addition here is a VIEW which provides a view on the
revisions table, but only the head of every tiddler. This can then
be used when doing searches. Without this results return for all
revisions which is clearly not the desired thing.

Managing the view is a little complex, so brought into _init_store().

Things that are still wrong:

* you cannot do field based search queries yet (like
  "ftitle:following jon") but that's just for want of a parser.
* tag based index_query searches are likely to have false positives
  because the query is a like...

This will be visited when more brain is available.