Update so that titles, tags and text are case sensitive.
This is a huge PITA because of the way mysql does collation.

This solution is a hack to adjust the table as it is being
created/mapped. On tiddlyspace.com the changes have already been made
by doing an alter table to change the collation on each column:
tiddler_title, tags, text.

In order to make those changes, the fulltext index has to be dropped,
the alterations done, then the fulltext index recreated.