Hack in a fix for a problem with doing index_query when there is a quote in the value.
This problem was found via tiddlyspace, where a tiddler with title
'foo"bar"baz' would not be found via the index. This is because
id:foo"bar"baz is not properly parsed.

A better long term solution is either to change the parser or
to change the quotes as they go through the parser.

We'll leave that for later.