ping a connection before each request to confirm that it is alive
This is a fallback to protect against "mysql server has gone away"
errors, which happen for reasons unknown. It seems like it might
be related to text and blob columns (wherein a request for a very
large blob works, but the connection is then closed without notifying
the client because some threshold has been crossed).

This checks every connection to see if it is live and creates another
if not.

The other option is to increase max_allowed_packet, but attempts thus
far at increasing it have been inconclusive, without removing fatal
errors, so this is required to be robust and flexible.

It will slightly slow things down.