Remove partial support for "NOT" in queries.
The partial support that was there would cause a 500 error whenever
'not' was used in seach query. Efforts to make 'NOT' actually work
lead to dead ends because of the structure of the fields tables. It
is possible that solutions exist, but my immediate concern is getting
rid of the 500 errors. When negation queries become a popular request
we'll investigate.