Add a magic search query optimization.
Some searches will return a lot of results, which we don't want,
so we them put a filter on them. This change optimizes for that
case by providing a way to declare the limit in the search query:

   q=starts _limit:10

Will get the 10 most recently modified tiddlers which match the
search query 'starts'.

This is most useful when generating feeds:

   q=modifier:cdent _limit:10

The leading '_' is there to not clobber existing searches that might
be looking for a limit field.