Add a failing test for NOT searches.
The SQL generated by a not query is incorrect.

For the search 'NOT tag:monkey' the search ought to look like this:

select distinct tiddler.title, tag.revision_number
  from tiddler, tag
    where 'monkey' not in (
      select tag from tag where tag.revision_number=tiddler.revision_number)
    and tiddler.revision_number=tag.revision_number;

but right now it looks more like this:

select distinct tiddler.title, tag.revision_number
  from tiddler, tag
    where tag.tag != 'monkey'
    and tiddler.revision_number=tag.revision_number;

The actual queries have joins in them, but the principle remains
the same.

Because of the way that "not" is handled in the search query parsing
and sql generation, it is hard to hork in the subquery shown above.

Or at least it is hard today when my brain is apparently missing a few
cylinders. So I leave it for the future.