Update to work with sqlalchemy 0.8.x
In that version when buiding a query, a Python None added to a
set of filters will result in a WHERE NULL clause, which means
you get no results.

The producer returns a None on some queries to indicate it
has fallen through, so we need to not push them into the
filter set.

@fnd will be proud to know that I was more able to effectively
debug this situation using `twanager interact` than with anything
else. Thanks.