fix #11 by _not_ aliasing the revision table
We don't need to because we have already effectively joined it
via the current_revision table. The upshot is that a search
which only inspects the revision table (not tags or fields)
ought to be somewhat less cpu intensive.

Unfortunately path between search query and result set is
so gnarly I don't feel confident with this change, despite
good tests. orz.